changed routing and template for gems-index
