+++ src/main/resources/templates/fragments.html
@@ -23,7 +23,7 @@
         </div>
         <div class="collapse navbar-collapse">
             <ul class="navbar-nav mr-auto">
-                <li class="nav-item"><a class="nav-link" th:href="@{/gems/index}">Browse Gems</a></li>
+                <li class="nav-item"><a class="nav-link" th:href="@{/gems-index}">Browse Gems</a></li>
                 <li class="nav-item"><a class="nav-link" th:href="@{/gems/add}"> Add Gem</a></li>
                 <li class="nav-item"><a class="nav-link" th:href="@{/reviews/add}">Add Gem Review</a></li>
             </ul>
