+++ src/main/resources/templates/gems-index.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html lang="en" xmlns:th="http://www.thymeleaf.org/">
+
+<head th:replace="fragments :: head">
+</head>
+<body>
+<header th:replace="fragments :: header"> </header>
+
+<div> All Gems:  </div>
+<div class="gem-index">
+    <table>
+        <tr th:each="gem : ${gems}">
+            <td>
+                <p>
+                    <a th:href="@{'gems/detail/' + ${gem.id}}" th:text="${gem.gemName}"></a>
+                </p>
+            </td>
+        </tr>
+
+    </table>
+</div>
+
+
+<footer th:replace="fragments :: footer"></footer>
+</body>
+</html>
